fix: move background color context closer together
