+++ packages/apps/human-app/frontend/src/shared/contexts/background/index.ts
@@ -1,0 +1,2 @@
+export * from './background';
+export * from './use-background-color-context';
